requirements.txt
pyyaml
requests
regex